+++ src/PersonPack/ComporatorsPack/PersonComparatorByChildren.java
@@ -1,6 +1,6 @@
-package Comporators;
+package PersonPack.ComporatorsPack;
 
-import Tree.Person;
+import PersonPack.Person;
 
 import java.util.Comparator;
 import java.util.Map;
