+++ src/UtilsPack/Savable.java
@@ -1,4 +1,4 @@
-package Utils;
+package UtilsPack;
 
 import java.io.Serializable;
